chore: attempts to fix integration tests
