+++ packages/core/src/components/navigation/Navbar/Navbar.tsx
@@ -1,4 +1,4 @@
-import { Suspense, useRef, useState } from 'react'
+import { useRef, useState } from 'react'
 
 import {
   Icon as UIIcon,
@@ -11,15 +11,13 @@ import {
   useUI,
 } from '@faststore/ui'
 
-import { mark } from 'src/sdk/tests/mark'
-
 import CartToggle from 'src/components/cart/CartToggle'
 import type { SearchInputRef } from 'src/components/search/SearchInput'
 import SearchInput from 'src/components/search/SearchInput'
 
+import { ButtonSignIn } from 'src/components/ui/Button'
 import Link from 'src/components/ui/Link'
 import Logo from 'src/components/ui/Logo'
-import { ButtonSignIn } from 'src/components/ui/Button'
 
 import type { NavbarProps as SectionNavbarProps } from '../../sections/Navbar'
 import NavbarLinks from '../NavbarLinks'
@@ -162,5 +160,4 @@ function Navbar({
   )
 }
 
-Navbar.displayName = 'Navbar'
-export default mark(Navbar)
+export default Navbar
